Remove unecessary git clones on Docker container
Our `docker/startup.sh` file cloned the `os-packager` repository and rebuilt
it. As we already have the code inside the container, I'm not sure why it's
useful. Looks like code that was required in the past, but not anymore. I
removed it and now we simply call "npm start" when running the container.

I also added a badge of the Docker build status to our README. We need a better
monitoring for the status of our builds, but while we don't have it, the badge
can be useful.