Remove authorization header before proxy request
We use Basic auth for the /pipelines endpoint, and an `authorization`
header may be present in the browser request for other endpoints to the
same domain. This commit ensures the authorization header is removed
from the proxy request to S3, otherwise the request fails with a
InvalidArgument: Unsupported Authorization Type error.